chore: refactor tasks that use open-ai compliant providers to extend a common class.
